ref (frontend): move month add modal
Move the month add modal to the entry screen directory.

Issue #25
