+++ web/src/app/screens/entry-screen/month-add-modal/month-add-modal.component.html
@@ -1,7 +1,9 @@
-<button type="button" class="btn btn-black tooltipbox" data-bs-toggle="modal" data-bs-target="#addMonthModal">
-  <span class="badge tooltiptext tooltiptext-left">Monat hinzufügen</span>
-  <i class="bi bi-calendar-plus"></i>
-</button>
+<div class="p-0 m-0 div-fixed-bottom-right d-flex justify-content-end">
+  <button type="button" class="btn btn-black tooltipbox" data-bs-toggle="modal" data-bs-target="#addMonthModal">
+    <span class="badge tooltiptext tooltiptext-left">Monat hinzufügen</span>
+    <i class="bi bi-calendar-plus"></i>
+  </button>
+</div>
 
 <div class="modal fade" id="addMonthModal" data-bs-backdrop="static" tabindex="-1" aria-labelledby="addMonthModalLabel"
      aria-hidden="true">
